+++ pkg/morph/contracts/methods.go
@@ -3,6 +3,7 @@ package fschaincontracts
 // Various methods of FS chain Container contract.
 const (
 	CreateContainerMethod         = "create"
+	CreateContainerV2Method       = "createV2"
 	RemoveContainerMethod         = "remove"
 	PutContainerEACLMethod        = "putEACL"
 	PutContainerReportMethod      = "putReport"
@@ -11,6 +12,7 @@ const (
 	GetTakenSpaceByUserMethod     = "getTakenSpaceByUser"
 	GetContainerQuotaMethod       = "containerQuota"
 	GetUserQuotaMethod            = "userQuota"
+	AddContainerStructsMethod     = "addStructs"
 )
 
 // CreateContainerParams are parameters of [CreateContainerMethod].